requirements.txt
python-gitlab==2.9.0
